Add first user test, move userRepository to classes direcory, create user class file
