+++ src/classes/User.js
@@ -1,0 +1,7 @@
+class User {
+     constructor() {
+
+     }
+}
+
+module.exports = User;
